20F

+++ src/main/java/com/at/business/IListaBusiness.java
@@ -6,11 +6,11 @@ import com.at.model.Lista;
 
 public interface IListaBusiness {
 
-	public Lista load(long id) throws BusinessException, NotFoundException;
+	public Lista load(int id) throws BusinessException, NotFoundException;
 
 	public Lista add(Lista lista) throws BusinessException;
 
-	public void delete(long id) throws BusinessException;
+	public void delete(int id) throws BusinessException;
 
 	public Lista update(Lista lista) throws BusinessException;
