+++ src/main/java/com/at/business/ITableroBusiness.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.at.model.Tablero;
 
 public interface ITableroBusiness {
-	public Tablero load(long id) throws BusinessException, NotFoundException;
+	public Tablero load(int id) throws BusinessException, NotFoundException;
 
 	public Tablero add(Tablero tablero) throws BusinessException;
 
-	public void delete(long id) throws BusinessException;
+	public void delete(int id) throws BusinessException;
 
 	public Tablero update(Tablero tablero) throws BusinessException;
